% Generated by roxygen2: do not edit by hand
\name{sqrtnumbeR}
\alias{sqrtnumbeR}
\title{Calculate the square of a number}
\usage{
sqrtnumbeR(number)
}
\arguments{
\item{number}{calculates the square of the number}
}
\value{
square
}
\description{
In this package a function is included to calculate the square of a number.
}
\examples{
sqrtnumbeR(3)

}
